+++ src/main/java/se/core/service/ResourceService.java
@@ -3,5 +3,5 @@ package se.core.service;
 import se.core.domain.Resource;
 
 public interface ResourceService {
-    Resource getResource(long id);
+    Resource getResource(Integer id);
 }
